start modularizing the code for easy readability
