+++ controllers/blog/api/filters.js
@@ -1,5 +1,5 @@
 const { countryGroup, articleGroup, extractGeoQuery } = require('./query')
-const { DB } = require('../../db')
+const { DB } = include('db/')
 
 exports.main = async kwargs => {
 	const conn = kwargs.connection ? kwargs.connection : DB.conn;
