+++ controllers/blog/blog_type/pdf.js
