+++ controllers/blog/scrapper/acclabs.js
@@ -1,9 +1,9 @@
 require("dotenv").config();
 const { Builder, By, until } = require("selenium-webdriver");
-const firefox = require("selenium-webdriver/firefox");
-const extractAndSaveData = require("./saveToDb");
-const { DB } = require("../../db");
+const extractAndSaveData = require("./save");
+const { DB } = include("db");
 const { checkUrlQuery } = require("./scrap-query");
+const setupWebDriver = require("../partial/webdriver");
 
 const acclab_publications = async () => {
   const baseurls = [
@@ -12,14 +12,14 @@ const acclab_publications = async () => {
     "https://www.undp.org/acceleratorlabs/blog",
   ];
 
-  // Set up the WebDriver
-  let options = new firefox.Options();
-    options.headless();
-  let driver = await new Builder()
-    .forBrowser("firefox")
-    .setFirefoxOptions(options)
-    .build();
-
+  // Setup WebDriver
+  let driver;
+  try {
+    driver = await setupWebDriver();
+  } catch (error) {
+    console.error("Error setting up WebDriver:", error);
+    return;
+  }
 await driver.manage().window().maximize();
   async function getButton() {
     let viewMoreButton;
